Fixed issues with code

+++ include/stdint.h
@@ -15,8 +15,16 @@ typedef unsigned short       uint16_t;
 typedef int                  int32_t;
 typedef unsigned int         uint32_t;
 
-typedef long long            int64_t;
-typedef unsigned long long   uint64_t;
+/* 32 bit on IPL32 systems*/
+//typedef long                 int64_t;
+//typedef unsigned long        uint64_t;
+
+/* 64 bit on IPL64 systems */
+typedef long                 int64_t;
+typedef unsigned long        uint64_t;
+
+//typedef long long            int64_t;
+//typedef unsigned long long   uint64_t;
 
 /* 7.18.1.4  Integer types capable of holding object pointers */
 typedef int intptr_t;
